added a superclass LoggedScreen
